Create tanssi-relay-encoder crate
