+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
 	"primitives/*",
 	"runtime/dancebox",
 	"runtime/flashbox",
+	"runtime/relay-encoder",
 ]
 resolver = "2"
 
@@ -71,6 +72,7 @@ flashbox-runtime = { path = "runtime/flashbox", default-features = false }
 manual-xcm-rpc = { path = "client/manual-xcm" }
 node-common = { path = "client/node-common" }
 runtime-common = { path = "runtime/common", default-features = false }
+tanssi-relay-encoder = { path = "runtime/relay-encoder", default-features = false }
 tc-consensus = { path = "client/consensus" }
 tp-author-noting-inherent = { path = "primitives/author-noting-inherent", default-features = false }
 tp-container-chain-genesis-data = { path = "primitives/container-chain-genesis-data", default-features = false }
