+++ runtime/relay-encoder/Cargo.toml
@@ -1,0 +1,56 @@
+[package]
+name = "tanssi-relay-encoder"
+authors = { workspace = true }
+description = "Allows to create encoded relay calls without depending on the relay runtime"
+edition = "2021"
+license = "GPL-3.0-only"
+version = "0.1.0"
+
+[package.metadata.docs.rs]
+targets = [ "x86_64-unknown-linux-gnu" ]
+
+[lints]
+workspace = true
+
+[dependencies]
+hex-literal = { workspace = true }
+parity-scale-codec = { workspace = true, features = [ "derive" ] }
+scale-info = { workspace = true, features = [ "derive" ] }
+
+# Substrate
+frame-support = { workspace = true }
+frame-system = { workspace = true }
+frame-try-runtime = { workspace = true, optional = true }
+pallet-balances = { workspace = true }
+
+# Cumulus
+sp-core = { workspace = true }
+sp-runtime = { workspace = true }
+sp-std = { workspace = true }
+
+cumulus-primitives-core = { workspace = true }
+
+[dev-dependencies]
+polkadot-runtime-parachains = { workspace = true }
+rococo-runtime = { workspace = true }
+rococo-runtime-constants = { workspace = true }
+
+[features]
+default = [
+	"std",
+]
+std = [
+	"cumulus-primitives-core/std",
+	"frame-support/std",
+	"frame-system/std",
+	"frame-try-runtime?/std",
+	"pallet-balances/std",
+	"parity-scale-codec/std",
+	"polkadot-runtime-parachains/std",
+	"rococo-runtime-constants/std",
+	"rococo-runtime/std",
+	"scale-info/std",
+	"sp-core/std",
+	"sp-runtime/std",
+	"sp-std/std",
+]
